+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/repository/BaseRankRepository.java
@@ -35,55 +35,94 @@ public class BaseRankRepository implements IBaseRankRepository {
         mRankClient = manager.getRankClien();
     }
 
+    /**
+     * @param size
+     * @return
+     */
     @Override
     public Observable<List<UserInfoBean>> getRankFollower(int size) {
-        return mRankClient.getRankFollower((long) TSListFragment.DEFAULT_PAGE_SIZE, size);
+        return mRankClient.getRankFollower(TSListFragment.DEFAULT_PAGE_SIZE, size);
     }
 
+    /**
+     * @param size
+     * @return
+     */
     @Override
     public Observable<List<UserInfoBean>> getRankRiches(int size) {
-        return mRankClient.getRankRiches((long) TSListFragment.DEFAULT_PAGE_SIZE, size);
+        return mRankClient.getRankRiches(TSListFragment.DEFAULT_PAGE_SIZE, size);
     }
 
+    /**
+     * @param size
+     * @return
+     */
     @Override
     public Observable<List<UserInfoBean>> getRankIncome(int size) {
-        return mRankClient.getRankIncome((long) TSListFragment.DEFAULT_PAGE_SIZE, size);
+        return mRankClient.getRankIncome(TSListFragment.DEFAULT_PAGE_SIZE, size);
     }
 
+    /**
+     * @param size
+     * @return
+     */
     @Override
     public Observable<List<UserInfoBean>> getRankCheckIn(int size) {
         SystemConfigBean systemConfigBean = SharePreferenceUtils.getObject(mContext, SharePreferenceTagConfig
                 .SHAREPREFERENCE_TAG_SYSTEM_BOOTSTRAPPERS);
         // 如果已经签到了，则不再展示签到
         if (systemConfigBean != null && systemConfigBean.isCheckin()) {
-            return mRankClient.getRankCheckIn((long) TSListFragment.DEFAULT_PAGE_SIZE, size);
-        }else {
+            return mRankClient.getRankCheckIn(TSListFragment.DEFAULT_PAGE_SIZE, size);
+        } else {
             return Observable.just(new ArrayList<>());
         }
     }
 
+    /**
+     * @param size
+     * @return
+     */
     @Override
     public Observable<List<UserInfoBean>> getRankQuestionExpert(int size) {
-        return mRankClient.getRankQuestionExpert((long) TSListFragment.DEFAULT_PAGE_SIZE, size);
+        return mRankClient.getRankQuestionExpert(TSListFragment.DEFAULT_PAGE_SIZE, size);
     }
 
+    /**
+     * @param size
+     * @return
+     */
     @Override
     public Observable<List<UserInfoBean>> getRankQuestionLikes(int size) {
-        return mRankClient.getRankQuestionLikes((long) TSListFragment.DEFAULT_PAGE_SIZE, size);
+        return mRankClient.getRankQuestionLikes(TSListFragment.DEFAULT_PAGE_SIZE, size);
     }
 
+    /**
+     * @param type
+     * @param size
+     * @return
+     */
     @Override
     public Observable<List<UserInfoBean>> getRankAnswer(String type, int size) {
-        return mRankClient.getRankAnswer(type, (long) TSListFragment.DEFAULT_PAGE_SIZE, size);
+        return mRankClient.getRankAnswer(type, TSListFragment.DEFAULT_PAGE_SIZE, size);
     }
 
+    /**
+     * @param type
+     * @param size
+     * @return
+     */
     @Override
     public Observable<List<UserInfoBean>> getRankDynamic(String type, int size) {
-        return mRankClient.getRankDynamic(type, (long) TSListFragment.DEFAULT_PAGE_SIZE, size);
+        return mRankClient.getRankDynamic(type, TSListFragment.DEFAULT_PAGE_SIZE, size);
     }
 
+    /**
+     * @param type
+     * @param size
+     * @return
+     */
     @Override
     public Observable<List<UserInfoBean>> getRankInfo(String type, int size) {
-        return mRankClient.getRankInfo(type, (long) TSListFragment.DEFAULT_PAGE_SIZE, size);
+        return mRankClient.getRankInfo(type, TSListFragment.DEFAULT_PAGE_SIZE, size);
     }
 }
